extract PromiseStatsWidget
