+++ src/components/Politician.vue
@@ -12,12 +12,7 @@
     </template>
     <template v-else>
     <h2>Promises by {{ politician.name }}</h2>
-    <el-card id="Politician_stats">
-      <b>Promise Statistics:</b>
-      <el-button v-for="stat in stats" :key="stat.value">
-        <b>{{ stat.value }}</b> {{ stat.number }}
-      </el-button>
-    </el-card>
+    <promise-stats-widget :promises="promises"/>
     <el-table
       :data="promises"
       border
@@ -80,22 +75,18 @@
 
 <script>
 import { getPolitician, listPoliticianPromises } from '@/api'
-import { generateStats, formatDate } from '@/utils'
+import { formatDate } from '@/utils'
 import LoadingSpinner from '@/components/shared/LoadingSpinner'
 import PoliticianDetails from '@/components/PoliticianDetails'
+import PromiseStatsWidget from '@/components/PromiseStatsWidget'
 
 export default {
   name: 'Politician',
-  components: { LoadingSpinner, PoliticianDetails },
+  components: { LoadingSpinner, PoliticianDetails, PromiseStatsWidget },
   data () {
     return {
       politician: 'loading',
       promises: 'loading'
-    }
-  },
-  computed: {
-    stats: function () {
-      return generateStats(this.promises)
     }
   },
   methods: {
